services/data/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
confluent-kafka==2.1.1
websockets==11.0.3
prometheus-client==0.17.0
circuitbreaker==1.4.0